+++ src/java/model/PricingRule.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 import model.base.BaseEntity;
  
 public class PricingRule extends BaseEntity {
-    private Long vehicleTypeId;
+    private int vehicleTypeId;
     private String bookingType;      // HOURLY, DAILY, DISTANCE, INNER_CITY, INTER_CITY
     private String tripDirection;    // ONE_WAY, ROUND_TRIP
     private BigDecimal pricePerKm;
@@ -13,15 +13,15 @@ public class PricingRule extends BaseEntity {
     private BigDecimal pricePerDay;
     private BigDecimal basePrice;
     private BigDecimal weekendMultiplier;
-    private Long updatedBy;
+    private int updatedBy;
     private Timestamp updatedAt;
  
     public PricingRule() {}
  
-    public PricingRule(Long vehicleTypeId, String bookingType, String tripDirection,
+    public PricingRule(int vehicleTypeId, String bookingType, String tripDirection,
                        BigDecimal pricePerKm, BigDecimal pricePerHour, BigDecimal pricePerDay,
                        BigDecimal basePrice, BigDecimal weekendMultiplier,
-                       Long updatedBy, Timestamp updatedAt) {
+                       int updatedBy, Timestamp updatedAt) {
         this.vehicleTypeId = vehicleTypeId;
         this.bookingType = bookingType;
         this.tripDirection = tripDirection;
@@ -34,8 +34,8 @@ public class PricingRule extends BaseEntity {
         this.updatedAt = updatedAt;
     }
  
-    public Long getVehicleTypeId() { return vehicleTypeId; }
-    public void setVehicleTypeId(Long vehicleTypeId) { this.vehicleTypeId = vehicleTypeId; }
+    public int getVehicleTypeId() { return vehicleTypeId; }
+    public void setVehicleTypeId(int vehicleTypeId) { this.vehicleTypeId = vehicleTypeId; }
  
     public String getBookingType() { return bookingType; }
     public void setBookingType(String bookingType) { this.bookingType = bookingType; }
@@ -58,8 +58,8 @@ public class PricingRule extends BaseEntity {
     public BigDecimal getWeekendMultiplier() { return weekendMultiplier; }
     public void setWeekendMultiplier(BigDecimal weekendMultiplier) { this.weekendMultiplier = weekendMultiplier; }
  
-    public Long getUpdatedBy() { return updatedBy; }
-    public void setUpdatedBy(Long updatedBy) { this.updatedBy = updatedBy; }
+    public int getUpdatedBy() { return updatedBy; }
+    public void setUpdatedBy(int updatedBy) { this.updatedBy = updatedBy; }
  
     public Timestamp getUpdatedAt() { return updatedAt; }
     public void setUpdatedAt(Timestamp updatedAt) { this.updatedAt = updatedAt; }
