+++ src/java/model/TripEventLog.java
@@ -4,14 +4,14 @@ import java.sql.Timestamp;
  
 public class TripEventLog {
     private Long id;
-    private Long bookingId;
+    private int bookingId;
     private String eventType;
     private String notes;
     private Timestamp eventTime;
  
     public TripEventLog() {}
  
-    public TripEventLog(Long bookingId, String eventType,
+    public TripEventLog(int bookingId, String eventType,
                         String notes, Timestamp eventTime) {
         this.bookingId = bookingId;
         this.eventType = eventType;
@@ -22,8 +22,8 @@ public class TripEventLog {
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
  
-    public Long getBookingId() { return bookingId; }
-    public void setBookingId(Long bookingId) { this.bookingId = bookingId; }
+    public int getBookingId() { return bookingId; }
+    public void setBookingId(int bookingId) { this.bookingId = bookingId; }
  
     public String getEventType() { return eventType; }
     public void setEventType(String eventType) { this.eventType = eventType; }
